Moving support functions to its own module
