+++ tests/data.py
@@ -5,7 +5,6 @@ import glob
 import pytest
 
 
-
 def get_sample_data(sample_path, file_type):
 
     if file_type == "12-00":
@@ -20,3 +19,31 @@ def get_sample_data(sample_path, file_type):
     print(f"Extracting: {output}")
     shutil.unpack_archive(output, sample_path)
     os.remove(output)
+
+
+@pytest.fixture
+def data_filenames():
+
+    home = os.path.expanduser("~")
+    sample_path = f"{home}/.lidarSuitrc/sample_data/"
+    file_type = "12-00"  # change to 6 beam in the future
+
+    if os.path.isdir(sample_path):
+
+        if os.path.isdir(f"{sample_path}{file_type}/"):
+            file_list = sorted(glob.glob(f"{sample_path}{file_type}/*.nc"))
+
+            if bool(file_list) == False:
+                get_sample_data(sample_path, file_type)
+                file_list = sorted(glob.glob(f"{sample_path}{file_type}/*.nc"))
+
+        else:
+            get_sample_data(sample_path, file_type)
+            file_list = sorted(glob.glob(f"{sample_path}{file_type}/*.nc"))
+
+    else:
+        os.makedirs(sample_path)
+        get_sample_data(sample_path, file_type)
+        file_list = sorted(glob.glob(f"{sample_path}{file_type}/*.nc"))
+
+    return file_list
