+++ tests/test_IO.py
@@ -1,40 +1,7 @@
-import os
-import gdown
-import shutil
-import glob
-import matplotlib.pyplot as plt
 import pytest
 
 import lidarSuit as lst
-from .data import get_sample_data
-
-
-@pytest.fixture
-def data_filenames():
-
-    home = os.path.expanduser("~")
-    sample_path = f"{home}/.lidarSuitrc/sample_data/"
-    file_type = "12-00"  # change to 6 beam in the future
-
-    if os.path.isdir(sample_path):
-
-        if os.path.isdir(f"{sample_path}{file_type}/"):
-            file_list = sorted(glob.glob(f"{sample_path}{file_type}/*.nc"))
-
-            if bool(file_list) == False:
-                get_sample_data(sample_path, file_type)
-                file_list = sorted(glob.glob(f"{sample_path}{file_type}/*.nc"))
-
-        else:
-            get_sample_data(sample_path, file_type)
-            file_list = sorted(glob.glob(f"{sample_path}{file_type}/*.nc"))
-
-    else:
-        os.makedirs(sample_path)
-        get_sample_data(sample_path, file_type)
-        file_list = sorted(glob.glob(f"{sample_path}{file_type}/*.nc"))
-
-    return file_list
+from .data import data_filenames, get_sample_data
 
 
 @pytest.fixture
